Organize mobx utils
